%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  as.CNA.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.CNA}
\alias{as.CNA}


 \title{Creating an 'as.CNA' object}

 \description{
  Creating an 'as.CNA' object.
 }

 \usage{as.CNA(genomdat.a, genomdat.b, chrom, maploc, normaldat.a, normaldat.b, genomdat.hetmatrix=NULL, normal.hetmatrix=NULL, sampleid=NULL, arraytype=NULL)}

 \arguments{
   \item{genomdat.a, genomdat.b}{}
   \item{chrom, maploc}{}
   \item{normaldat.a, normaldat.b}{}
   \item{genomdat.hetmatrix}{}
   \item{normal.hetmatrix}{}
   \item{sampleid}{}
   \item{arraytype}{}
 }

 \value{
   Returns a \code{\link[base]{list}} structure of class \code{as.CNA}.
 }

 \author{Adam Olshen <olshena@biostat.ucsf.edu>}

 \seealso{
   \code{\link{as.segment}}().
 }

