%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ASCBS-package}
\alias{ASCBS-package}
\alias{ASCBS}
\docType{package}

\title{Package ASCBS}




\description{
  Allele-specific segmentation and calling using Circular Binary Segmentation.

  This package should be considered to be in an alpha phase.
  You should expect the API to be changing over time.
  This is a private package.
}

\author{
  Adam Olshen <olshena@biostat.ucsf.edu>
}

\section{License}{
  file LICENSE
}
\keyword{package}
